lu_acc not set on reject, removed old timestamp copy in session_info, fixed si.sql integers

+++ si.sql
@@ -6,7 +6,7 @@ CREATE TABLE session_info (
   domain tinyint NOT NULL,		-- Communication domain (Circuit Switched=0, Packet Switched=1)
   mcc smallint NOT NULL,		-- Mobile country code where the transaction was recorded
   mnc smallint NOT NULL,		-- Mobile network code where the transaction was recorded
-  lac smallint NOT NULL,		-- Location area code where the transaction was recorded
+  lac int NOT NULL,			-- Location area code where the transaction was recorded
   cid int NOT NULL,			-- Cell ID where the transaction was recorded
   arfcn int DEFAULT NULL,		-- ARFCN of current MCC/MNC/LAC/CID
   psc tinyint DEFAULT NULL,		-- Primary scrambling code of current cell 
@@ -51,7 +51,7 @@ CREATE TABLE session_info (
   lu_rej_cause tinyint DEFAULT NULL,	-- Location update was rejected with this cause
   lu_mcc smallint DEFAULT NULL,		-- MCC of previous user location
   lu_mnc smallint DEFAULT NULL,		-- MNC of previous user location
-  lu_lac smallint DEFAULT NULL,		-- LAC of previous user location
+  lu_lac int DEFAULT NULL,		-- LAC of previous user location
   t_raupd tinyint DEFAULT NULL,		-- Transaction contains a routing area update
   t_attach tinyint DEFAULT NULL,	-- Transaction contains a GPRS attach
   att_acc tinyint DEFAULT NULL,		-- GPRS attach was accepted
